+++ NEWS.md
@@ -1,3 +1,24 @@
+0.2.0 (2016-08-11)
+------------------
+* Adds Gravity Model
+
+0.1.0 (2016-06-29)
+------------------
+* Adds Spatial Markov function
+* Adds Spacial interpolation function
+* Adds `CDB_pyAgg (columns Numeric[])` helper function
+* Adds Segmentation Functions
+
+0.0.4 (2016-06-20)
+------------------
+* Remove cartodb extension dependency from tests
+* Declare all correct dependencies with correct versions in setup.py
+
+0.0.3 (2016-06-16)
+------------------
+* Adds new functions: kmeans, weighted centroids.
+* Replaces moran functions with new areas of interest naming.
+
 0.0.2 (2016-03-16)
 ------------------
 * New versioning approach using per-version Python virtual environments